+++ resources/views/components/no-records-found.blade.php
@@ -1,0 +1,5 @@
+<tr>
+    <td colspan="999" class="text-center fst-italic text-muted">
+        {{__('No Records Found')}}
+    </td>
+</tr>